backend/requirements.txt
fastapi[standard]
uvicorn[standard]
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
pydantic[email]
openpyxl
